refactor: rework variable creation and constraint wrapper
Major rework in the variable creation and handling within the
constraint wrapper. fixed variables are now created seperately and the
solver does not know about them anymore. the rule conditions are now
checked within the solver. so moving closer to a expert system fixed
bugs in the particle initialization stage of intermediate edges fixed
bugs in the cparity rule and condition function and unit test
introduced None as a not defined value for the cparity