refactor!: change conservation rule interface (ComPWA/expertsystem#253)
* refactor!: change `Rule` interface
  - Change rule arguments from list of dict based objects to list of
custom data objects. Custom meaning either single type or composite
dataclasses.
  - Define a set of quantum numbers and its types (via `NewType`) for edge
and nodes separately. See `EdgeQuantumNumbers` and `NodeQuantumNumbers`
  - State rule requirements in terms of type annotations of the arguments.
This is now a code requirement when writing rules.
  - Remove composite spin quantum numbers. Magnitude and projection can now
 only be used separately.

* fix: fix `FullPropagator` case handling (temporary)
Note: This is a temporary fix. In the future a correct solutions should
refactor the propagation code and introduce and + or operations for
rules (most importantly is the or operation).

* chore: complete conservation rule type annotations

* docs: complete docstrings for conservation rules

Co-authored-by: Remco de Boer <redeboer@gmx.com>