refactor!: make solving module graph independent (ComPWA/expertsystem#401)
* refactor!: Change interface of Solver, to accept a new data structure
QNProblemSet as an argument. Also introduce a new data structure
QNResult which is now the new return type of the Solver interface. This
removes the StateTransitionGraph from the module.

* refactor!: Move Result to the reaction module to keep the interface for
find_solutions of the STM stable.

* refactor!: Define new InitialFacts and ProblemSet data structures which
are part of the new UI. Change prepare_graphs to create_problem_sets,
that now creates and returns ProblemSets. These can be fed to the
solving module.

* refactor!: Change CSPSolver init to accept allowed particles in the form
of quantum number dictionaries. This simplifies the code and removes the
coupling to the Particle module. The quantum number dictionaries have to
use None as a value for undefined Parities.

* test: test particle to edge properties conversion