feat: reduce edge qn domains to allowed states  (ComPWA/expertsystem#406)
* fix: correct merging of qn solution with particle

Fixes a bug that was introduced with PR ComPWA/expertsystem#401. It is related merging
solutions with particle states when more than one intermediate edge is
present and resulted in no found solutions. 

Also enable the test which can detect such kind of problems again (was
previously disabled because of long runtime).

* feat: reduce edge qn domains to allowed states

In case allowed intermediate states are specified, reduce the edge
quantum numbers for intermediate edges to match the given allowed
states.
This can dramatically reduce the runtime, when intermediate states are
explicitly specified.

* ci: temporarily turn off mac os builds